fix: Fix paths Refs #400
