+++ src/main/resources/application-test.properties
@@ -1,29 +1,21 @@
-spring.application.name=infantem-test
-
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=test
-spring.datasource.password=test
-spring.datasource.init.schema-locations=classpath:/test_data.sql
-
-
+database=mysql
+spring.datasource.url=jdbc:mysql://localhost:3306/infantem_db
+spring.datasource.username=spring_user
+spring.datasource.password=StrongPassword123!
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.sql.init.mode=always
+spring.jpa.defer-datasource-initialization=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.defer-datasource-initialization=true
 
-
-infantem.jwt.secret=test_secret
+infantem.jwt.secret=este_es_un_secretillo1234!!
+infantem.jwt.timeout=86400000
 infantem.jwt.timeout=7776000000
 
-logging.level.root=INFO
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
-
-spring.jpa.properties.hibernate.cache.use_second_level_cache=false
-spring.jpa.properties.hibernate.cache.use_query_cache=false
 
 
 spring.mail.host=smtp.gmail.com
@@ -44,3 +36,4 @@ spring.servlet.multipart.enabled=true
 spring.servlet.multipart.max-file-size=10MB
 spring.servlet.multipart.max-request-size=10MB
 spring.servlet.multipart.resolve-lazily=true
+
